Flash page write tested

+++ FreeRTOS_PoC_F4/App/Inc/protocol.h
@@ -47,7 +47,7 @@ typedef struct{
 
 typedef struct{
 	uint32_t address;
-	uint8_t size;
+	uint16_t size;	// because we can read up to 256 bytes (0x100)
 	uint8_t data[FLASH_PAGE_SIZE];
 } WRITE_FLASH_IN;
